Moving endpoint from /forgotPassword to /users.forgotPassword
