+++ packages/rocketchat-api/server/v1/users.js
@@ -393,3 +393,18 @@ RocketChat.API.v1.addRoute('user.roles', { authRequired: true }, {
 		}));
 	}
 });
+
+RocketChat.API.v1.addRoute('users.forgotPassword', { authRequired: false }, {
+	post() {
+		const { email } = this.bodyParams;
+		if (!email) {
+			return RocketChat.API.v1.failure('The \'email\' param is required');
+		}
+
+		const emailSent = Meteor.call('sendForgotPasswordEmail', email);
+		if (emailSent) {
+			return RocketChat.API.v1.success();
+		}
+		return RocketChat.API.v1.failure('User not found');
+	}
+});
