+++ tests/end-to-end/api/01-users.js
@@ -589,4 +589,32 @@ describe('[Users]', function() {
 				.end(done);
 		});
 	});
+
+	describe('[/users.forgotPassword]', () => {
+		it('should send email to user (return success), when is a valid email', (done) => {
+			request.post(api('users.forgotPassword'))
+				.send({
+					email: adminEmail
+				})
+				.expect('Content-Type', 'application/json')
+				.expect(200)
+				.expect((res) => {
+					expect(res.body).to.have.property('success', true);
+				})
+				.end(done);
+		});
+
+		it('should not send email to user(return error), when is a invalid email', (done) => {
+			request.post(api('users.forgotPassword'))
+				.send({
+					email: 'invalidEmail'
+				})
+				.expect('Content-Type', 'application/json')
+				.expect(400)
+				.expect((res) => {
+					expect(res.body).to.have.property('success', false);
+				})
+				.end(done);
+		});
+	});
 });
